+++ app/src/main/java/com/canopas/yourspace/ui/flow/journey/timeline/JourneyTimelineViewModel.kt
@@ -98,7 +98,7 @@ class JourneyTimelineViewModel @Inject constructor(
         try {
             val from = _state.value.selectedTimeFrom
             val to = _state.value.selectedTimeTo
-            val lastJourneyTime = allJourneys.minOfOrNull { it.update_at!! }
+            val lastJourneyTime = allJourneys.minOfOrNull { it.updated_at!! }
 
             val locations = if (loadMore) {
                 journeyService.getMoreJourneyHistory(userId, lastJourneyTime)
@@ -108,7 +108,7 @@ class JourneyTimelineViewModel @Inject constructor(
 
             val filteredLocations = locations.filter {
                 (it.created_at?.let { created -> created in from..to } ?: false) ||
-                    (it.update_at?.let { updated -> updated in from..to } ?: false)
+                    (it.updated_at?.let { updated -> updated in from..to } ?: false)
             }
 
             val locationJourneys = (allJourneys + filteredLocations).groupByDate()
@@ -177,7 +177,7 @@ class JourneyTimelineViewModel @Inject constructor(
 
     private fun List<LocationJourney>.groupByDate(): Map<Long, List<LocationJourney>> {
         val journeys = this.distinctBy { it.id }
-            .sortedByDescending { it.update_at!! }
+            .sortedByDescending { it.updated_at!! }
 
         val groupedItems = mutableMapOf<Long, MutableList<LocationJourney>>()
