+++ data/src/main/java/com/canopas/yourspace/data/service/location/ApiJourneyService.kt
@@ -1,9 +1,10 @@
 package com.canopas.yourspace.data.service.location
 
-import com.canopas.yourspace.data.models.location.EncryptedJourneyRoute
 import com.canopas.yourspace.data.models.location.EncryptedLocationJourney
 import com.canopas.yourspace.data.models.location.JourneyRoute
 import com.canopas.yourspace.data.models.location.LocationJourney
+import com.canopas.yourspace.data.models.location.toDecryptedLocationJourney
+import com.canopas.yourspace.data.models.location.toEncryptedLocationJourney
 import com.canopas.yourspace.data.models.space.GroupKeysDoc
 import com.canopas.yourspace.data.models.user.ApiUser
 import com.canopas.yourspace.data.storage.UserPreferences
@@ -14,7 +15,6 @@ import com.canopas.yourspace.data.utils.Config.FIRESTORE_COLLECTION_SPACE_GROUP_
 import com.canopas.yourspace.data.utils.Config.FIRESTORE_COLLECTION_SPACE_MEMBERS
 import com.canopas.yourspace.data.utils.EphemeralECDHUtils
 import com.canopas.yourspace.data.utils.PrivateKeyUtils
-import com.google.firebase.firestore.Blob
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.Query
 import com.google.firebase.firestore.toObject
@@ -81,7 +81,7 @@ class ApiJourneyService @Inject constructor(
             val groupCipher = GroupCipher(bufferedSenderKeyStore, groupAddress)
             return Pair(distributionMessage, groupCipher)
         } catch (e: Exception) {
-            Timber.e(e, "Error processing group session for spaceId=$spaceId, userId=$userId")
+            Timber.e(e, "Error processing group session")
             return null
         }
     }
@@ -149,7 +149,7 @@ class ApiJourneyService @Inject constructor(
                 route_duration = routeDuration,
                 routes = routes,
                 created_at = createdAt ?: System.currentTimeMillis(),
-                update_at = updateAt ?: System.currentTimeMillis()
+                updated_at = updateAt ?: System.currentTimeMillis()
             )
 
             val docRef = spaceMemberJourneyRef(spaceId, userId).document(journey.id)
@@ -160,93 +160,6 @@ class ApiJourneyService @Inject constructor(
 
             docRef.set(encryptedJourney).await()
         }
-    }
-
-    private fun LocationJourney.toEncryptedLocationJourney(
-        groupCipher: GroupCipher,
-        distributionId: UUID
-    ): EncryptedLocationJourney {
-        val encryptedFromLat = groupCipher.encrypt(
-            distributionId,
-            from_latitude.toString().toByteArray(Charsets.UTF_8)
-        )
-        val encryptedFromLong = groupCipher.encrypt(
-            distributionId,
-            from_longitude.toString().toByteArray(Charsets.UTF_8)
-        )
-        val encryptedToLat = to_latitude?.let {
-            groupCipher.encrypt(
-                distributionId,
-                it.toString().toByteArray(Charsets.UTF_8)
-            )
-        }
-        val encryptedToLong = to_longitude?.let {
-            groupCipher.encrypt(
-                distributionId,
-                it.toString().toByteArray(Charsets.UTF_8)
-            )
-        }
-
-        val encryptedRoutes = routes.map {
-            EncryptedJourneyRoute(
-                encrypted_latitude = Blob.fromBytes(
-                    groupCipher.encrypt(
-                        distributionId,
-                        it.latitude.toString().toByteArray(Charsets.UTF_8)
-                    ).serialize()
-                ),
-                encrypted_longitude = Blob.fromBytes(
-                    groupCipher.encrypt(
-                        distributionId,
-                        it.longitude.toString().toByteArray(Charsets.UTF_8)
-                    ).serialize()
-                )
-            )
-        }
-
-        return EncryptedLocationJourney(
-            id = id,
-            user_id = user_id,
-            encrypted_from_latitude = Blob.fromBytes(encryptedFromLat.serialize()),
-            encrypted_from_longitude = Blob.fromBytes(encryptedFromLong.serialize()),
-            encrypted_to_latitude = encryptedToLat?.let { Blob.fromBytes(it.serialize()) },
-            encrypted_to_longitude = encryptedToLong?.let { Blob.fromBytes(it.serialize()) },
-            route_distance = route_distance,
-            route_duration = route_duration,
-            encrypted_routes = encryptedRoutes,
-            created_at = created_at,
-            updated_at = update_at
-        )
-    }
-
-    private fun EncryptedLocationJourney.toDecryptedLocationJourney(groupCipher: GroupCipher): LocationJourney {
-        val decryptedFromLat = groupCipher.decrypt(encrypted_from_latitude.toBytes())
-        val decryptedFromLong = groupCipher.decrypt(encrypted_from_longitude.toBytes())
-        val decryptedToLat = encrypted_to_latitude?.let { groupCipher.decrypt(it.toBytes()) }
-        val decryptedToLong = encrypted_to_longitude?.let { groupCipher.decrypt(it.toBytes()) }
-
-        val decryptedRoutes = encrypted_routes.map {
-            JourneyRoute(
-                latitude = groupCipher.decrypt(it.encrypted_latitude.toBytes())
-                    .toString(Charsets.UTF_8).toDouble(),
-                longitude = groupCipher.decrypt(it.encrypted_longitude.toBytes())
-                    .toString(Charsets.UTF_8).toDouble()
-            )
-        }
-
-        return LocationJourney(
-            id = id,
-            user_id = user_id,
-            from_latitude = decryptedFromLat.toString(Charsets.UTF_8).toDouble(),
-            from_longitude = decryptedFromLong.toString(Charsets.UTF_8).toDouble(),
-            to_latitude = decryptedToLat?.toString(Charsets.UTF_8)?.toDouble(),
-            to_longitude = decryptedToLong?.toString(Charsets.UTF_8)?.toDouble(),
-            route_distance = route_distance,
-            route_duration = route_duration,
-            routes = decryptedRoutes,
-            created_at = created_at,
-            update_at = updated_at
-        )
     }
 
     suspend fun updateLastLocationJourney(userId: String, journey: LocationJourney) {
